requirements.txt
appdirs==1.4.0
attrs==16.3.0
cffi==1.9.1
constantly==15.1.0
cryptography==1.7.2
cssselect==1.0.1
idna==2.2
incremental==16.10.1
lxml==3.7.2
packaging==16.8
parsel==1.1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.17
PyDispatcher==2.0.5
pymongo==3.4.0
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
queuelib==1.4.2
Scrapy==1.3.0
service-identity==16.0.0
six==1.10.0
Twisted==16.6.0
w3lib==1.16.0
zope.interface==4.3.3
